Start the downloader service as a foreground service on Oreo and up
If the user sends the app to the background while they're waiting for
the APK expansion to download, calls to the alarm receiver might cause
the app to attempt to start the service from the background.  Under Oreo
and up, this will cause the app to be killed unless we explicitly start
the service as a foreground service.